refactor(fetch_htlm): param 'crawl_timeout' data type chaged to int
In 'fetch_html' data type of param 'crawl_timeout' changed from float to int

+++ graphcrawl_ai/extraction/html_from_url.py
@@ -9,7 +9,7 @@ logging.basicConfig(
 
 class FetchError(Exception): pass
 
-def fetch_html(url: str = "", crawl_timeout: float = 30, crawl_retry: int = 3) -> str:
+def fetch_html(url: str = "", crawl_timeout: int = 30, crawl_retry: int = 3) -> str:
     """Download the raw HTML content from a given web address.
 
     This function attempts to retrieve the webpage content by performing
